Chore: Various Package Updates (#138)
* Change: Update package versions for InfiniBlazor and Lucide

Updated InfiniBlazor packages to version 0.1.0-preview.21 and Lucide to 0.26.509. This ensures compatibility with the latest features and bug fixes of these dependencies.

* Feat: Update dependencies and add new package

Added `CodeOfChaos.Extensions` to project references and updated multiple `CodeOfChaos` package versions to `0.66.1`. These changes ensure compatibility and include the latest improvements and fixes for the dependencies.

* Change: Update InfiniBlazor package versions to preview.22

Updated the InfiniBlazor packages to version 0.1.0-preview.22 across all related dependencies. This ensures compatibility with the latest features and bug fixes in the preview release.

* Change: Upgrade package versions for dependencies

Updated FastEndpoints to version 6.1.0 and TUnit to version 0.19.143, along with their related packages. This ensures compatibility with the latest features and fixes provided by these libraries.

* Change: Update package versions to latest stable releases

* Feat: Add InfiniBlazor.Toasting package and update versions

* Feat: Upgrade TailwindCSS to version 4.1.6 and update dependencies

* Change: Update workflow to use correct solution